The color attributes were lost in scrollback buffer since the 256-color change.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@38222 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/HistoryBuffer.cpp
@@ -79,7 +79,7 @@ HistoryBuffer::GetTerminalLineAt(int32 index, TerminalLine* buffer) const
 	int32 charCount = 0;
 	const char* chars = line->Chars();
 	buffer->length = 0;
-	uint16 attributes = 0;
+	uint32 attributes = 0;
 	AttributesRun* attributesRun = line->AttributesRuns();
 	int32 attributesRunCount = line->attributesRunCount;
 	int32 nextAttributesAt = attributesRunCount > 0
@@ -128,7 +128,7 @@ HistoryBuffer::AddLine(const TerminalLine* line)
 {
 //debug_printf("HistoryBuffer::AddLine(%p): length: %d\n", line, line->length);
 	// determine the amount of memory we need for the line
-	uint16 attributes = 0;
+	uint32 attributes = 0;
 	int32 attributesRuns = 0;
 	int32 byteLength = 0;
 	for (int32 i = 0; i < line->length; i++) {
