+++ src/apps/terminal/TerminalLine.h
@@ -30,7 +30,7 @@ struct TerminalLine {
 
 
 struct AttributesRun {
-	uint16	attributes;
+	uint32	attributes;
 	uint16	offset;			// character offset
 	uint16	length;			// length of the run in characters
 };
